Fix A32-on-A64 for Apple Silicon
Unpatch is called from InvalidateBasicBlocks, which has temporarily
enabled writing, and thus disabled execution. However, the fast dispatch
lookup code is dynamically generated just like the rest of the JITed
output and so also is currently not executable. Thus we must temporarily
disable writing, and thus re-enable execution, whilst we run the lookup
function, otherwise we will fault on jumping to it.